wordsdb

+++ my_dict/tasks.py
@@ -1,5 +1,5 @@
 from django.utils import timezone
-from .models import WotdEn, Profile
+from .models import WotdEn, Profile, WordDb
 from bs4 import BeautifulSoup
 import requests
 import lxml
@@ -16,11 +16,13 @@ def reset_quiz():
 
 def word_of_the_day():
     try:
-        all_wotds = WotdEn.objects.all()
-        for wotd in all_wotds:
-            wotd.delete()
+        if WotdEn.objects.all():
+            all_wotds = WotdEn.objects.all()
+            for wotd in all_wotds:
+                wotd.delete()
     except:
         pass
+
     req_mer = requests.get('https://www.merriam-webster.com/', headers={"User-Agent": "Mozilla/5.0"}).content
     req_ox = requests.get('https://www.oxfordlearnersdictionaries.com/', headers={"User-Agent": "Mozilla/5.0"}).content
     req_dict = requests.get('https://www.dictionary.com/', headers={"User-Agent": "Mozilla/5.0"}).content
@@ -39,12 +41,20 @@ def word_of_the_day():
     tur_camb = []
     if search_word(word_mer) is not None:
         tur_mer = search_word(word_mer)
+    else:
+        tur_mer = [None, None]
     if search_word(word_ox) is not None:
         tur_ox = search_word(word_ox)
+    else:
+        tur_ox = [None, None]
     if search_word(word_dict) is not None:
         tur_dict = search_word(word_dict)
+    else:
+        tur_dict = [None, None]
     if search_word(word_camb) is not None:
         tur_camb = search_word(word_camb)
+    else:
+        tur_camb = [None, None]
 
     create_mer = WotdEn.objects.create(english=word_mer, audio=tur_mer[1], website="Merriam")
     create_mer.save()
@@ -55,18 +65,22 @@ def word_of_the_day():
     create_camb = WotdEn.objects.create(english=word_camb, audio=tur_camb[1], website="Cambridge")
     create_camb.save()
 
-    for tr in tur_mer[0]:
-        create_tr = create_mer.turkish.create(turkish=tr)
-        create_tr.save()
-    for tr in tur_ox[0]:
-        create_tr = create_ox.turkish.create(turkish=tr)
-        create_tr.save()
-    for tr in tur_dict[0]:
-        create_tr = create_dict.turkish.create(turkish=tr)
-        create_tr.save()
-    for tr in tur_camb[0]:
-        create_tr = create_camb.turkish.create(turkish=tr)
-        create_tr.save()
+    if tur_mer[0] is not None:
+        for tr in tur_mer[0]:
+            create_tr = create_mer.turkish.create(turkish=tr)
+            create_tr.save()
+    if tur_ox[0] is not None:
+        for tr in tur_ox[0]:
+            create_tr = create_ox.turkish.create(turkish=tr)
+            create_tr.save()
+    if tur_dict[0] is not None:
+        for tr in tur_dict[0]:
+            create_tr = create_dict.turkish.create(turkish=tr)
+            create_tr.save()
+    if tur_camb[0] is not None:
+        for tr in tur_camb[0]:
+            create_tr = create_camb.turkish.create(turkish=tr)
+            create_tr.save()
 
 
 def search_word(word):
@@ -75,28 +89,56 @@ def search_word(word):
     url = pre_url.format(str(plus_added_search))
     html = requests.get(url).content
     soup = BeautifulSoup(html, 'lxml')
-    table = soup.find('table')
-    try:
-        rows = table.find_all('tr')[1:]
-        audio_and_tr = []
-        tr_list = []
 
+    try:
+        table = soup.find('table')
+        rows = table.find_all('tr')[1:]
+        audio_and_tr = [None, None]
+        tr_list = []
         for row in rows:
             if not row.attrs:
                 tds = row.find_all('td')
                 tr = tds[3].text.strip()
                 tr_list.append(tr)
+        if tr_list:
+            audio_and_tr[0] = tr_list[:3]
+        else:
+            audio_and_tr[0] = None
 
-        audio_and_tr.append(tr_list[:3])
-
-        if soup.find('audio', {'id': 'turengVoiceENTRENus'}):
+        try:
             if soup.find('audio', {'id': 'turengVoiceENTRENus'}).find('source'):
                 audio = soup.find('audio', {'id': 'turengVoiceENTRENus'}).find('source')['src']
-                audio_and_tr.append(audio)
-        else:
-            audio_and_tr.append(None)
+                audio_and_tr[1] = audio
+            else:
+                audio_and_tr[1] = None
+        except:
+            pass
 
-        return audio_and_tr[0], audio_and_tr[1]
-
-    except AttributeError:
+        return audio_and_tr
+    except:
         return None
+
+
+def add_words2_db():
+    url = "https://twinword-word-graph-dictionary.p.rapidapi.com/association/"
+    headers = {
+        'x-rapidapi-key': "c5cbdceae4msh740bd52ea8d5e8bp1f6881jsn7c651aa39638",
+        'x-rapidapi-host': "twinword-word-graph-dictionary.p.rapidapi.com"
+    }
+    github = "https://github.com/first20hours/google-10000-english/blob/master/google-10000-english-usa-no-swears-long.txt"
+    html_tur = requests.get(github).content
+    soup_tur = BeautifulSoup(html_tur, 'lxml')
+    tds = soup_tur.find_all('td', class_='blob-code')[:100]
+    for td in tds:
+        response = requests.request("GET", url, headers=headers, params={'entry': td.text})
+        en = response.json()['response']
+        obj = WordDb.objects.create(english=en)
+        obj.save()
+
+
+
+
+
+
+
+
